chore: centralize looseAttributePrefixes sanitization, add changeset and docs note
- Move sanitizeLooseAttributePrefixes into svelte2tsx internalHelpers and
  apply it inside convertHtmlxToJsx so all entry points are guarded
- Reuse the shared helper in language-server, svelte-check and
  typescript-plugin (closes the unsanitized path in the plugin)
- Add changeset for the new feature
- Document that svelteOptions is not picked up through tsconfig extends
