+++ packages/language-server/src/plugins/typescript-go/features/DiagnosticsProvider.ts
@@ -37,7 +37,6 @@ import { DiagnosticsProvider } from '../../interfaces';
 import {
     DocumentSnapshot,
     JSOrTSDocumentSnapshot,
-    sanitizeLooseAttributePrefixes,
     SvelteDocumentSnapshot,
     SvelteSnapshotOptions
 } from '../../typescript/DocumentSnapshot';
@@ -120,9 +119,10 @@ export class SvelteCheckTSGoDiagnosticsProvider implements DiagnosticsProvider {
         if (this.projectConfig?.raw.svelteOptions?.namespace) {
             this.snapshotOptions.typingsNamespace = this.projectConfig.raw.svelteOptions.namespace;
         }
-        this.snapshotOptions.looseAttributePrefixes = sanitizeLooseAttributePrefixes(
-            this.projectConfig?.raw.svelteOptions?.looseAttributePrefixes
-        );
+        this.snapshotOptions.looseAttributePrefixes =
+            internalHelpers.sanitizeLooseAttributePrefixes(
+                this.projectConfig?.raw.svelteOptions?.looseAttributePrefixes
+            );
         this.pendingConfigLoading = configLoader.loadConfigs(path.dirname(tsconfigPath));
         this.writeVirtualTsconfig();
     }
