+++ packages/language-server/src/plugins/typescript/DocumentSnapshot.ts
@@ -111,20 +111,6 @@ export interface SvelteSnapshotOptions {
     looseAttributePrefixes?: string[];
 }
 
-/**
- * Validates the `svelteOptions.looseAttributePrefixes` tsconfig value, which is
- * unchecked user input.
- */
-export function sanitizeLooseAttributePrefixes(prefixes: unknown): string[] | undefined {
-    if (!Array.isArray(prefixes)) {
-        return undefined;
-    }
-    const sanitized = prefixes.filter(
-        (prefix): prefix is string => typeof prefix === 'string' && prefix.length > 0
-    );
-    return sanitized.length > 0 ? sanitized : undefined;
-}
-
 const ambientPathPattern = /node_modules[\/\\]svelte[\/\\]types[\/\\]ambient\.d\.ts$/;
 const svelteTypesPattern = /node_modules[\/\\]svelte[\/\\]types[\/\\]index\.d\.ts$/;
 const shimsPattern =
