+++ packages/svelte2tsx/index.d.ts
@@ -180,6 +180,7 @@ export const internalHelpers: {
     ) => { text: string; addedCode: InternalHelpers.AddedCode[] } | undefined,
     toVirtualPos: (pos: number, addedCode: InternalHelpers.AddedCode[]) => number,
     toOriginalPos: (pos: number, addedCode: InternalHelpers.AddedCode[]) => {pos: number; inGenerated: boolean},
+    sanitizeLooseAttributePrefixes: (prefixes: unknown) => string[] | undefined,
     findExports: (_ts: typeof ts, source: ts.SourceFile, isTsFile: boolean) => Map<
         string,
         | {
