+++ packages/svelte2tsx/src/htmlxtojsx_v2/index.ts
@@ -11,6 +11,7 @@ import {
     DeclarationTag
 } from '../interfaces';
 import { parseHtmlx } from '../utils/htmlxparser';
+import { sanitizeLooseAttributePrefixes } from '../helpers/looseAttributePrefixes';
 import { handleActionDirective } from './nodes/Action';
 import { handleAnimateDirective } from './nodes/Animation';
 import { handleAttribute } from './nodes/Attribute';
@@ -107,6 +108,9 @@ export function convertHtmlxToJsx(
     } = { svelte5Plus: false }
 ): TemplateProcessResult {
     options.typingsNamespace = options.typingsNamespace || 'svelteHTML';
+    // The value may come straight from an unvalidated tsconfig, so sanitize it here
+    // to guard all entry points at once
+    options.looseAttributePrefixes = sanitizeLooseAttributePrefixes(options.looseAttributePrefixes);
     const preserveAttributeCase = options.namespace === 'foreign';
     const emitJsDoc = options.emitJsDoc ?? false;
     const isTsFile = options.isTsFile ?? false;
